requirements.txt
gigachat
psutil
unstructured[pdf]
python-poppler
networkx
unstructured
langchain
langchain_community
pdfplumber
chromadb
beautifulsoup4
python-dotenv
logging
requests
sqlalchemy
pandas
openpyxl
telebot